// React/app/GerAdv_TS/OperadorGruposAgenda/Interfaces/interface.OperadorGruposAgenda.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IOperadorGruposAgenda {
// 202501251
    id: number;
 
	operador: number,
	sqlwhere: string,
	nome: string,
}

export interface IOperadorGruposAgendaFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IOperadorGruposAgendaIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}